Remove different delegates for different message types
