+++ Sources/Protocols/MessageInputBarDelegate.swift
@@ -23,7 +23,6 @@
  */
 
 import Foundation
-import UIKit
 
 /// MessageInputBarDelegate is a protocol that can recieve notifications from the MessageInputBar
 public protocol MessageInputBarDelegate: AnyObject {
