+++ Sources/Protocols/MessagesDisplayDelegate.swift
@@ -23,51 +23,12 @@
  */
 
 import Foundation
-
-/// A protocol used by the `MessagesViewController` to customize the appearance of a `TextMessageCell`.
-public protocol TextMessageDisplayDelegate: AnyObject {
-
-    /// Specifies the color of the text for a `TextMessageCell`.
-    ///
-    /// - Parameters:
-    ///   - message: A `MessageType` with a `MessageData` case of `.text` or `.attributedText` to which the color will apply.
-    ///   - indexPath: The `IndexPath` of the cell.
-    ///   - messagesCollectionView: The `MessagesCollectionView` in which this cell will be displayed.
-    ///
-    /// The default value returned by this method is determined by the messages `Sender`:
-    ///
-    /// Current Sender: UIColor.white
-    ///
-    /// All other Senders: UIColor.darkText
-    func textColor(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> UIColor
-
-    /// Specifies the `DetectorType`s to check for the `MessageType`'s text against.
-    ///
-    /// - Parameters:
-    ///   - message: A `MessageType` with a `MessageData` case of `.text` or `.attributedText` to which the detectors will apply.
-    ///   - indexPath: The `IndexPath` of the cell.
-    ///   - messagesCollectionView: The `MessagesCollectionView` in which this cell will be displayed.
-    ///
-    /// The default value returned by this method is all available detector types.
-    func enabledDetectors(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> [DetectorType]
-
-}
-
-public extension TextMessageDisplayDelegate {
-
-    func textColor(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> UIColor {
-        guard let dataSource = messagesCollectionView.messagesDataSource else { return .darkText }
-        return dataSource.isFromCurrentSender(message: message) ? .white : .darkText
-    }
-
-    func enabledDetectors(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> [DetectorType] {
-        return []
-    }
-
-}
+import MapKit
 
 /// A protocol used by the `MessagesViewController` to customize the appearance of a `MessagesCollectionViewCell`.
 public protocol MessagesDisplayDelegate: AnyObject {
+
+    // MARK: - All Messages
 
     /// Specifies the `MessageStyle` to be used for a `MessageContainerView`.
     ///
@@ -123,9 +84,73 @@ public protocol MessagesDisplayDelegate: AnyObject {
     /// The default value returned by this method is a `MessageFooterView`.
     func messageFooterView(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> MessageFooterView
 
+    // MARK: - Text Messages
+
+    /// Specifies the color of the text for a `TextMessageCell`.
+    ///
+    /// - Parameters:
+    ///   - message: A `MessageType` with a `MessageData` case of `.text` or `.attributedText` to which the color will apply.
+    ///   - indexPath: The `IndexPath` of the cell.
+    ///   - messagesCollectionView: The `MessagesCollectionView` in which this cell will be displayed.
+    ///
+    /// The default value returned by this method is determined by the messages `Sender`:
+    ///
+    /// Current Sender: UIColor.white
+    ///
+    /// All other Senders: UIColor.darkText
+    func textColor(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> UIColor
+
+    /// Specifies the `DetectorType`s to check for the `MessageType`'s text against.
+    ///
+    /// - Parameters:
+    ///   - message: A `MessageType` with a `MessageData` case of `.text` or `.attributedText` to which the detectors will apply.
+    ///   - indexPath: The `IndexPath` of the cell.
+    ///   - messagesCollectionView: The `MessagesCollectionView` in which this cell will be displayed.
+    ///
+    /// The default value returned by this method is all available detector types.
+    func enabledDetectors(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> [DetectorType]
+
+    // MARK: - Location Messages
+
+    /// Ask the delegate for a LocationMessageSnapshotOptions instance to customize the MapView on the given message
+    ///
+    /// - Parameters:
+    ///   - message: The location message to be customized
+    ///   - indexPath: Message's index path
+    ///   - messagesCollectionView: The collection view requesting the information
+    /// - Returns: Your LocationMessageSnapshotOptions instance with the options to customize map style
+    func snapshotOptionsForLocation(message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> LocationMessageSnapshotOptions
+
+    /// Ask the delegate for a custom MKAnnotationView to show on the given message.
+    /// You can return nil if you don't want to show any annotation.
+    ///
+    /// default: MKPinAnnotationView
+    ///
+    /// - Parameters:
+    ///   - message: The location message with the annotation to customize
+    ///   - indexPath: Message's index path
+    ///   - messageCollectionView: The collection view requesting the information
+    /// - Returns: Your customized MKAnnotationView or nil to not show any.
+    func annotationViewForLocation(message: MessageType, at indexPath: IndexPath, in messageCollectionView: MessagesCollectionView) -> MKAnnotationView?
+
+    /// Ask the delegate for a custom animation block to run when whe map screenshot is ready to be displaied in the given location message
+    /// The animation block is called with the image view to be animated. You can animate it with CoreAnimation, UIView.animate or any library you prefer.
+    ///
+    /// default: nil
+    ///
+    /// - Parameters:
+    ///   - message: The location message with the map to animate
+    ///   - indexPath: Message's index path
+    ///   - messagesCollectionView: The collection view requesting the information
+    /// - Returns: Your customized animation block.
+    func animationBlockForLocation(message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> ((UIImageView) -> Void)?
+
+
 }
 
 public extension MessagesDisplayDelegate {
+
+    // MARK: - All Messages Defaults
 
     func messageStyle(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> MessageStyle {
         return .bubble
@@ -160,6 +185,31 @@ public extension MessagesDisplayDelegate {
 
     func messageFooterView(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> MessageFooterView {
         return messagesCollectionView.dequeueReusableFooterView(MessageFooterView.self, for: indexPath)
+    }
+
+    // MARK: - Text Messages Defaults
+
+    func textColor(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> UIColor {
+        guard let dataSource = messagesCollectionView.messagesDataSource else { return .darkText }
+        return dataSource.isFromCurrentSender(message: message) ? .white : .darkText
+    }
+
+    func enabledDetectors(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> [DetectorType] {
+        return []
+    }
+
+    // MARK: - Location Messages Defaults
+
+    func snapshotOptionsForLocation(message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> LocationMessageSnapshotOptions {
+        return LocationMessageSnapshotOptions()
+    }
+
+    func annotationViewForLocation(message: MessageType, at indexPath: IndexPath, in messageCollectionView: MessagesCollectionView) -> MKAnnotationView? {
+        return MKPinAnnotationView(annotation: nil, reuseIdentifier: nil)
+    }
+
+    func animationBlockForLocation(message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> ((UIImageView) -> Void)? {
+        return nil
     }
 
 }
