+++ Sources/Protocols/MessagesLayoutDelegate.swift
@@ -23,24 +23,13 @@
  */
 
 import Foundation
+import AVFoundation
 
 /// A protocol used by the `MessagesCollectionViewFlowLayout` object to determine
 /// the size and layout of a `MessageCollectionViewCell` and its contents.
 public protocol MessagesLayoutDelegate: AnyObject {
 
-    /// Specifies the insets for the text rect of the `MessageLabel` in a `TextMessageCell`.
-    ///
-    /// - Parameters:
-    ///   - message: A `MessageType` with a `MessageData` case of `.text` or `.attributedText` to which these insets will apply.
-    ///   - indexPath: The `IndexPath` of the cell.
-    ///   - messagesCollectionView: The `MessagesCollectionView` in which this cell will be displayed.
-    ///
-    /// The default value returned by this method is determined by the messages `Sender`:
-    ///
-    /// Current Sender: `UIEdgeInsets(top: 7, left: 14, bottom: 7, right: 18)`
-    ///
-    /// All other Senders: `UIEdgeInsets(top: 7, left: 18, bottom: 7, right: 14)`
-    func messageLabelInset(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> UIEdgeInsets
+    // MARK: - All Messages
 
     /// Specifies the padding around the `MessageContainerView` in a `MessageCollectionViewCell`.
     ///
@@ -125,18 +114,75 @@ public protocol MessagesLayoutDelegate: AnyObject {
     /// The default value returned by this method is a size of `GGSize.zero`.
     func footerViewSize(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> CGSize
 
+    // MARK: - Text Messages
+
+    /// Specifies the insets for the text rect of the `MessageLabel` in a `TextMessageCell`.
+    ///
+    /// - Parameters:
+    ///   - message: A `MessageType` with a `MessageData` case of `.text` or `.attributedText` to which these insets will apply.
+    ///   - indexPath: The `IndexPath` of the cell.
+    ///   - messagesCollectionView: The `MessagesCollectionView` in which this cell will be displayed.
+    ///
+    /// The default value returned by this method is determined by the messages `Sender`:
+    ///
+    /// Current Sender: `UIEdgeInsets(top: 7, left: 14, bottom: 7, right: 18)`
+    ///
+    /// All other Senders: `UIEdgeInsets(top: 7, left: 18, bottom: 7, right: 14)`
+    func messageLabelInset(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> UIEdgeInsets
+
+    // MARK: - Media Messages
+
+    /// Specifies the width for a `MessageContainerView`.
+    ///
+    /// - Parameters:
+    ///   - message: The `MessageType` that will be displayed by this cell.
+    ///   - indexPath: The `IndexPath` of the cell.
+    ///   - maxWidth: The max available width for the `MessageContainerView` respecting the cell's other content.
+    ///   - messagesCollectionView: The `MessagesCollectionView` in which this cell will be displayed.
+    ///
+    /// The default value returned by this method is the `maxWidth`.
+    func widthForMedia(message: MessageType, at indexPath: IndexPath, with maxWidth: CGFloat, in messagesCollectionView: MessagesCollectionView) -> CGFloat
+
+    /// Specifies the height for a `MessageContainerView`.
+    ///
+    /// - Parameters:
+    ///   - message: The `MessageType` that will be displayed by this cell.
+    ///   - indexPath: The `IndexPath` of the cell.
+    ///   - maxWidth: The max available width for the `MessageContainerView` respecting the cell's other content.
+    ///   - messagesCollectionView: The `MessagesCollectionView` in which this cell will be displayed.
+    ///
+    /// The default value returned by this method uses `AVMakeRect(aspectRatio:insideRect:)` with a bounding
+    /// rect using the `maxWidth` and `.greatestFiniteMagnitude` for the height.
+    func heightForMedia(message: MessageType, at indexPath: IndexPath, with maxWidth: CGFloat, in messagesCollectionView: MessagesCollectionView) -> CGFloat
+
+
+    // MARK: - Location Messages
+
+    /// Specifies the width for a `MessageContainerView`.
+    ///
+    /// - Parameters:
+    ///   - message: The `MessageType` that will be displayed by this cell.
+    ///   - indexPath: The `IndexPath` of the cell.
+    ///   - maxWidth: The max available width for the `MessageContainerView` respecting the cell's other content.
+    ///   - messagesCollectionView: The `MessagesCollectionView` in which this cell will be displayed.
+    ///
+    /// The default value returned by this method is the `maxWidth`.
+    func widthForLocation(message: MessageType, at indexPath: IndexPath, with maxWidth: CGFloat, in messagesCollectionView: MessagesCollectionView) -> CGFloat
+
+    /// Specifies the height for a `MessageContainerView`.
+    ///
+    /// - Parameters:
+    ///   - message: The `MessageType` that will be displayed by this cell.
+    ///   - indexPath: The `IndexPath` of the cell.
+    ///   - maxWidth: The max available width for the `MessageContainerView` respecting the cell's other content.
+    ///   - messagesCollectionView: The `MessagesCollectionView` in which this cell will be displayed.
+    func heightForLocation(message: MessageType, at indexPath: IndexPath, with maxWidth: CGFloat, in messagesCollectionView: MessagesCollectionView) -> CGFloat
+
 }
 
 public extension MessagesLayoutDelegate {
 
-    func messageLabelInset(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> UIEdgeInsets {
-        guard let dataSource = messagesCollectionView.messagesDataSource else { return .zero }
-        if dataSource.isFromCurrentSender(message: message) {
-            return UIEdgeInsets(top: 7, left: 14, bottom: 7, right: 18)
-        } else {
-            return UIEdgeInsets(top: 7, left: 18, bottom: 7, right: 14)
-        }
-    }
+    // MARK: - All Messages Defaults
 
     func messagePadding(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> UIEdgeInsets {
         guard let dataSource = messagesCollectionView.messagesDataSource else { return .zero }
@@ -173,6 +219,39 @@ public extension MessagesLayoutDelegate {
 
     func footerViewSize(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> CGSize {
         return .zero
+    }
+
+    // MARK: - Text Messages Defaults
+
+    func messageLabelInset(for message: MessageType, at indexPath: IndexPath, in messagesCollectionView: MessagesCollectionView) -> UIEdgeInsets {
+        guard let dataSource = messagesCollectionView.messagesDataSource else { return .zero }
+        if dataSource.isFromCurrentSender(message: message) {
+            return UIEdgeInsets(top: 7, left: 14, bottom: 7, right: 18)
+        } else {
+            return UIEdgeInsets(top: 7, left: 18, bottom: 7, right: 14)
+        }
+    }
+
+    // MARK: - Media Messages Defaults
+
+    func widthForMedia(message: MessageType, at indexPath: IndexPath, with maxWidth: CGFloat, in messagesCollectionView: MessagesCollectionView) -> CGFloat {
+        return maxWidth
+    }
+
+    func heightForMedia(message: MessageType, at indexPath: IndexPath, with maxWidth: CGFloat, in messagesCollectionView: MessagesCollectionView) -> CGFloat {
+        switch message.data {
+        case .photo(let image), .video(_, let image):
+            let boundingRect = CGRect(origin: .zero, size: CGSize(width: maxWidth, height: .greatestFiniteMagnitude))
+            return AVMakeRect(aspectRatio: image.size, insideRect: boundingRect).height
+        default:
+            return 0
+        }
+    }
+
+    // MARK: - Location Messages Defaults
+
+    func widthForLocation(message: MessageType, at indexPath: IndexPath, with maxWidth: CGFloat, in messagesCollectionView: MessagesCollectionView) -> CGFloat {
+        return maxWidth
     }
 
 }
